requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.0
cron-descriptor==1.4.0
dispatcher==1.0
Django==4.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
future==0.18.3
hiredis==2.3.2
idna==3.6
inflection==0.5.1
kombu==5.3.4
magic-filter==1.0.12
multidict==6.0.4
packaging==23.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyTelegramBotAPI==4.14.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
serializers==0.2.4
six==1.16.0
sqlparse==0.4.4
telebot==0.0.5
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
